Adding an options class for BlobDB and also Impl class.
Summary: 1. Options Class.

2. Impl class BlobDBImpl
3. made BlobDB::Open

Test Plan:

Reviewers:

Subscribers:

Tasks:

Blame Revision:
