+++ utilities/blob_db/blob_db.h
@@ -10,9 +10,11 @@
 #include <string>
 #include "rocksdb/db.h"
 #include "rocksdb/status.h"
+#include "rocksdb/utilities/stackable_db.h"
 
 namespace rocksdb {
-// EXPERIMENAL ONLY
+
+
 // A wrapped database which puts values of KV pairs in a separate log
 // and store location to the log in the underlying DB.
 // It lacks lots of importatant functionalities, e.g. DB restarts,
@@ -20,5 +22,51 @@ namespace rocksdb {
 //
 // The factory needs to be moved to include/rocksdb/utilities to allow
 // users to use blob DB.
-extern Status NewBlobDB(Options options, std::string dbname, DB** blob_db);
+
+struct BlobDBOptions {
+
+  // name of the directory under main db, where blobs will be stored.
+  std::string blob_dir;
+
+  // whether the blob_dir path is relative or absolute.
+  bool path_relative;
+
+  // Is the blob db mindful of ttl and eviction is done on TTL.
+  bool has_ttl;
+
+  // a new bucket is opened, for ttl_range. So if ttl_range is 600seconds
+  // (10 minutes), and the first bucket starts at 1471542000
+  // then the blob buckets will be
+  // first bucket is 1471542000 - 1471542600
+  // second bucket is 1471542600 - 1471543200
+  // and so on
+  uint64_t ttl_range;
+
+  // at what size will the blobs be stored in separate log rather than
+  // inline
+  uint64_t min_blob_size;
+
+  // at what bytes will the blob files be synced to blob log.
+  uint64_t bytes_per_sync;
+};
+
+class BlobDB : public StackableDB {
+ public:
+  using rocksdb::StackableDB::Put;
+  Status Put(const WriteOptions& options, const Slice& key,
+             const Slice& value) override  = 0;
+
+  using rocksdb::StackableDB::Get;
+  Status Get(const ReadOptions& options, const Slice& key,
+             std::string* value) override = 0;
+
+  static Status Open(const Options& options, const BlobDBOptions& bdb_options,
+      const std::string& dbname, BlobDB** blob_db);
+
+protected:
+
+  explicit BlobDB(DB* db);
+
+};
+
 }  // namespace rocksdb
